[refactor] routes 파일 구조
