+++ ongi/lib/screens/helpers/alarm_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:table_calendar/table_calendar.dart';
 import 'package:ongi/widgets/bottom_nav_bar/bottom_nav_bar_with_alarm.dart';
-import 'package:ongi/screens/helpors/home_screen.dart';
-import 'package:ongi/screens/helpors/schedule_screen.dart';
-import 'package:ongi/screens/helpors/settings_screen.dart';
+import 'package:ongi/screens/helpers/home_screen.dart';
+import 'package:ongi/screens/helpers/schedule_screen.dart';
+import 'package:ongi/screens/helpers/settings_screen.dart';
 
 class AlarmScreen extends StatefulWidget {
   const AlarmScreen({Key? key}) : super(key: key);
