+++ ongi/lib/test_main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'screens/helpors/home_screen.dart'; // 또는 emergency_alert_screen.dart
+import 'screens/helpers/home_screen.dart'; // 또는 emergency_alert_screen.dart
 
 void main() {
   runApp(const MyApp());
